requirements.txt
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pyyaml>=6.0
